Refactor lix-game packages into a scope
Eliminates the confusing intertwingling between the various packages,
which made it easy to lose track of who's defining what, where
certain settings should be `override`n, etc.

Also added hashes for the assets, both with and without PNG32
conversion. Fixes #1. (We'll see if the PNG32 hash is consistent
across platforms or not.)