minivmac: Remove 'renamed' wrappers; rename for real
Something went wrong when I was porting that code from my overlay to
here, and it kept getting an infinite recursion error (or a segfault).
I tried to fix it, but eventually gave up and ripped that code out, and
let Nix rebuild Mini vMac whenever the name is changed. It's small, it's
fine.

Also, I added a suffix to non-Plus builds of Mini vMac by default, so
renaming shouldn't be necessary as often now anyway.